refactor(roles): split RolesService into CommunityRolesService and InstanceRolesService
Splits the 1,461-line RolesService into two focused services, following the
sibling-services precedent in backend/src/messages/:

- CommunityRolesService: community-scoped role CRUD, well-known role lookups
  (admin/moderator/member), and user-role assignment within a community.
- InstanceRolesService (implements OnModuleInit): instance-level role CRUD,
  the Community Creator / User Manager / Invite Manager well-known roles, and
  the default-instance-roles bootstrap that runs on module init.

No facade — roles.service.ts and roles.service.spec.ts are deleted. Method
bodies moved verbatim (only the bumpNowOrDefer call-path, established in the
prior commit, and DI wiring changed). No cross-calls exist between the two
services' methods, so no service was injected into the other.

Updates every consumer (roles.controller.ts now injects both services;
user/community/moderation/membership services take CommunityRolesService;
onboarding.service.ts takes both) and their specs' TestBed mock targets, plus
the e2e test helper that re-seeds default instance roles.

Splits roles.service.spec.ts's 122 its across the two new spec files
(69 community + 53 instance), recreating both the "Permission cache epoch
bumps" and "Deferred epoch bumps inside caller-owned transactions" describe
blocks in each file per the mutations they own.
