+++ backend/src/roles/roles.module.ts
@@ -1,5 +1,6 @@
 import { Module } from '@nestjs/common';
-import { RolesService } from './roles.service';
+import { CommunityRolesService } from './community-roles.service';
+import { InstanceRolesService } from './instance-roles.service';
 import { PermissionsService } from './permissions.service';
 import { PermissionsCacheService } from './permissions-cache.service';
 import { DatabaseModule } from '@/database/database.module';
@@ -8,8 +9,18 @@ import { RolesController } from './roles.controller';
 
 @Module({
   imports: [DatabaseModule, RedisModule],
-  providers: [RolesService, PermissionsService, PermissionsCacheService],
-  exports: [RolesService, PermissionsService, PermissionsCacheService],
+  providers: [
+    CommunityRolesService,
+    InstanceRolesService,
+    PermissionsService,
+    PermissionsCacheService,
+  ],
+  exports: [
+    CommunityRolesService,
+    InstanceRolesService,
+    PermissionsService,
+    PermissionsCacheService,
+  ],
   controllers: [RolesController],
 })
 export class RolesModule {}
